Implement truncate_slug/3 to limit length of slug.
truncate_slug/3 takes a slug like the ones that get generated by slugify/2 and a maximum_length and enforces that maximum length on the slug.

I bumped the Elixir version since I used to_charlist which is the successor of the softdeprecated to_char_list, but is not available in Elixir 1.0.

Fixes #11