Add :hard option to truncate_slug/3
Change the last argument of truncate_slug/3 to an options kwlist, which let's the user configure the separator and/or if the truncation should be done with or without breaking inside a slugged word (truncate hard).

I'm not a 100% sure, if the wording for that option (:hard) is good. Feedback welcome.